pyproject.toml
[tool.poetry]
name = "flaskproject"
version = "0.1.0"
description = "Coursework for Flask framework"
authors = ["Nikita Bulatov <nikita_bulatov@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
python-dotenv = "^1.0.0"
flask-login = "^0.6.2"
flask-migrate = "^4.0.4"
psycopg2 = "^2.9.5"
wtforms = "^3.0.1"
email-validator = "^1.3.1"
flask-wtf = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
